refactor(genai): rename filenames
